[NFC] Minor fixes to improve debugging.
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/151053
Tested-by: Shkurko, Dmitry V <dmitry.v.shkurko@intel.com>
Old/split llvm commit ID: 2aec35b6e583882594fb4a32f8df81d4d83b5eba

+++ llvm/Intel_OptionalComponents/DTrans/lib/DTransCommon.cpp
@@ -28,6 +28,11 @@ using namespace llvm;
 static cl::opt<unsigned> DTransMemLayoutLevel("dtrans-mem-layout-level",
                                               cl::init(2), cl::ReallyHidden);
 
+// Delete fields transformation.
+static cl::opt<bool> EnableDeleteFields("enable-dtrans-deletefield",
+                                        cl::init(true), cl::Hidden,
+                                        cl::desc("Enable DTrans delete field"));
+
 #if !defined(NDEBUG) || defined(LLVM_ENABLE_DUMP)
 // Valid values: early -> dump before early DTrans passes
 //               late -> dump before late DTrans passes
@@ -54,25 +59,32 @@ static cl::opt<bool>
     EnableResolveTypes("enable-resolve-types", cl::init(true),
                         cl::Hidden,
                         cl::desc("Enable pre-dtrans type resolution"));
+
+// SOA-to-AOS transformation.
+static cl::opt<bool> EnableSOAToAOS("enable-dtrans-soatoaos", cl::init(false),
+                                    cl::Hidden,
+                                    cl::desc("Enable DTrans SOAToAOS"));
+
 #else
 
 #define hasDumpModuleBeforeDTransValue(x) (false)
 constexpr bool EnablePaddedPtrProp = false;
 constexpr bool EnableResolveTypes = false;
+constexpr bool EnableSOAToAOS = false;
 
 #endif // !defined(NDEBUG) || defined(LLVM_ENABLE_DUMP)
 
 void llvm::initializeDTransPasses(PassRegistry &PR) {
   initializeDTransAnalysisWrapperPass(PR);
-  initializeDTransAOSToSOAWrapperPass(PR);
-  initializeDTransDeleteFieldWrapperPass(PR);
   initializeDTransPaddedMallocWrapperPass(PR);
   initializePaddedPtrPropWrapperPass(PR);
-  initializeDTransReorderFieldsWrapperPass(PR);
   initializeDTransResolveTypesWrapperPass(PR);
+  initializeDTransSOAToAOSWrapperPass(PR);
+  initializeDTransDeleteFieldWrapperPass(PR);
+  initializeDTransReorderFieldsWrapperPass(PR);
+  initializeDTransAOSToSOAWrapperPass(PR);
   initializeDTransEliminateROFieldAccessWrapperPass(PR);
   initializeDTransDynCloneWrapperPass(PR);
-  initializeDTransSOAToAOSWrapperPass(PR);
 
 #if !INTEL_PRODUCT_RELEASE
   initializeDTransOptBaseTestWrapperPass(PR);
@@ -87,12 +99,14 @@ void llvm::addDTransPasses(ModulePassManager &MPM) {
   if (EnableResolveTypes)
     MPM.addPass(dtrans::ResolveTypesPass());
 
-  MPM.addPass(dtrans::DeleteFieldPass());
+  if (EnableSOAToAOS)
+    MPM.addPass(dtrans::SOAToAOSPass());
+  if (EnableDeleteFields)
+    MPM.addPass(dtrans::DeleteFieldPass());
   MPM.addPass(dtrans::ReorderFieldsPass());
   MPM.addPass(dtrans::AOSToSOAPass());
   MPM.addPass(dtrans::EliminateROFieldAccessPass());
   MPM.addPass(dtrans::DynClonePass());
-  MPM.addPass(dtrans::SOAToAOSPass());
 }
 
 void llvm::addDTransLegacyPasses(legacy::PassManagerBase &PM) {
@@ -103,12 +117,14 @@ void llvm::addDTransLegacyPasses(legacy::PassManagerBase &PM) {
   if (EnableResolveTypes)
     PM.add(createDTransResolveTypesWrapperPass());
 
-  PM.add(createDTransDeleteFieldWrapperPass());
+  if (EnableSOAToAOS)
+    PM.add(createDTransSOAToAOSWrapperPass());
+  if (EnableDeleteFields)
+    PM.add(createDTransDeleteFieldWrapperPass());
   PM.add(createDTransReorderFieldsWrapperPass());
   PM.add(createDTransAOSToSOAWrapperPass());
   PM.add(createDTransEliminateROFieldAccessWrapperPass());
   PM.add(createDTransDynCloneWrapperPass());
-  PM.add(createDTransSOAToAOSWrapperPass());
 }
 
 void llvm::addLateDTransPasses(ModulePassManager &MPM) {
